requirements.txt
watchdog==3.0.0
GitPython==3.1.37
PyYAML==6.0.1
PyGithub==2.1.1
schedule==1.2.0 